fullstack-chatbot/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.11
chromadb>=0.4.0
faiss-cpu>=1.7.0
openai>=1.0.0
sentence-transformers>=2.2.0
unstructured>=0.10.0
docx2txt>=0.8
pypdf>=3.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
